requirements.txt
psycopg2-binary  
django-heroku  
gunicorn  
django-storages 
boto3 
django-import-export
reportlab
PyPDF2
pandas
google-auth
google-cloud-storage
